requirements.txt
Django==4.2.2
django-cors==0.1
django-cors-headers==4.0.0
django-jazzmin==2.6.0
djangorestframework==3.14.0
duckduckgo-search==3.8.3
httpcore==0.17.3
httpx==0.24.1
hyperframe==6.0.1
numpy==1.25.2
psycopg2==2.9.6
PyJWT==2.7.0